[update]chatのモデルの分離、階層の変更 #36
