+++ opt/chat/models/room_message.py
@@ -4,16 +4,7 @@ import sys
 
 sys.path.append('../')
 from accounts.models.user import User
-
-
-class Room(models.Model):
-    name = models.CharField(max_length=100)
-    member_count = models.IntegerField(default=0)
-
-
-class RoomMember(models.Model):
-    room = models.ForeignKey(Room, on_delete=models.CASCADE)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
+from chat.models.room import Room
 
 
 class RoomMessage(models.Model):
